Split subreddit into submodules and sub-submodules
