+++ praw/models/reddit/subreddit/flair/__init__.py
@@ -1,0 +1,6 @@
+"""Container for SubredditFlair and templates."""
+# flake8: noqa
+from .flair import SubredditFlair
+from .link import SubredditLinkFlairTemplates
+from .redditor import SubredditRedditorFlairTemplates
+from .templates import SubredditFlairTemplates
